Moved components that are rendered from  toolbox into a folder also cleaned up some old imports that we didn't need and some more cleanup
