+++ app/components/ElementDragPreview.js
@@ -1,7 +1,5 @@
 import React, { Component } from 'react';
-import TextElement from './ElementComponents/TextElement';
-import PhotoElement from './ElementComponents/PhotoElement'
-import VideoElement from './ElementComponents/VideoElement'
+import whatTypeElementToRender from './utils/whatTypeElementToRender'
 
 const styles = {
   display: 'inline-block',
@@ -10,40 +8,11 @@ const styles = {
 
 export default class ElementDragPreview extends Component {
 
-  constructor (props) {
-    super(props)
-
-    this.whatElementToRender = this.whatElementToRender.bind(this)
-  }
-
-  whatElementToRender() {
-    //switch case based off of the element type,
-    //returns the proper element
-    const { text, id, source, type, size } = this.props;
-
-    switch (this.props.type) {
-
-      case "textBox":
-        return (
-          <TextElement text={text} id={id} size={size} type={type} {...this.props}/>
-        );
-      case "photo":
-        return (
-          <PhotoElement id={id} size={size} source={source}/>
-        );
-      case "video":
-        return (
-          <VideoElement id={id} size={size} source={source}/>
-        );
-    }
-  }
-
-
   render() {
 
     return (
       <div style={styles}>
-        {this.whatElementToRender()}
+        {whatTypeElementToRender(this.props)}
       </div>
     );
   }
