+++ app/reducers/elements.js
@@ -1,14 +1,19 @@
 
 const initialState = {
   textBox: {
+    /*
     1: { top: 20, left: 80, size: 'large', text: 'My vacay memories', background: '#51b6ea', color: '#000000' },
-    2: { top: 100, left: 120, size: 'small', text: 'note to self', background: '#51b6ea', color: '#000000' }
+     2: { top: 100, left: 120, size: 'small', text: 'note to self', background: '#51b6ea', color: '#000000' }
+     */
   },
   photo: {
-    1: {top: 200, left: 200, size: "small", source: "http://placehold.it/300/09f/fff.png"}
+    /* 1: {top: 200, left: 200, size: "small", source: "http://placehold.it/300/09f/fff.png"}
+    */
   },
   video: {
+    /*
     3: {top: 300, left: 300, size: "small", source: ""}
+    */
   }
 }
 
@@ -38,8 +43,6 @@ const reducer = (state = initialState, action) => {
 
     case 'ADD_A_VIDEO':
       id = Object.keys(action.video)[0]
-      // console.log("INSIDE", Object.keys(action.video)[0])
-      // console.log("VIDEO IN REDUCER ", action.video, Object.keys(action.photo)[0])
       newState.video[id] = action.video[id]
       break;
